feat: support decimal values in calculateCanvasSize

+++ src/calculateCanvasSize/calculateCanvasSize.ts
@@ -1,6 +1,6 @@
 export default function calculateCanvasSize(length: string, width: string) {
-  const parsedLength = parseInt(length);
-  const parsedWidth = parseInt(width);
+  const parsedLength = parseFloat(length);
+  const parsedWidth = parseFloat(width);
 
   if (isNaN(parsedLength) || isNaN(parsedWidth)) {
     throw new Error("Invalid input: length and width must be numbers");
